Move PullAdapter to extensions-management and api modules (#1289)
